requirements.txt
aenum==3.1.16
aiohttp==3.9.5
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Deprecated==1.2.18
Django==5.2.2
fonttools==4.58.4
frozenlist==1.7.0
future==1.0.0
idna==3.10
joblib==1.5.1
kiwisolver==1.4.8
line-bot-sdk==3.11.0
Markdown==3.8.2
matplotlib==3.10.3
multidict==6.6.3
numpy==2.3.1
packaging==25.0
pillow==11.2.1
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.31.0
scikit-learn==1.7.0
scipy==1.16.0
setuptools==78.1.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.1
